@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,9 +27,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
   intl: ^0.17.0
-  flutter_svg: ^0.19.1
+  flutter_svg: ^0.22.0
+  firebase_core: ^1.3.0
+  cloud_firestore: ^2.2.2
 
 dev_dependencies:
   flutter_test:
