@@ -3,10 +3,10 @@ import 'package:workhours/components/campotextogeral.dart';
 import 'package:workhours/constants.dart';
 
 class Camposenha extends StatelessWidget {
-  final ValueChanged<String> onChanged;
-  final TextEditingController controller;
+  final ValueChanged<String> ?onChanged;
+  final TextEditingController ?controller;
   const Camposenha({
-    Key key,
+    Key ?key,
     this.onChanged,
     this.controller
   }) : super(key: key);
