@@ -1,4 +1,5 @@
 import 'dart:ui';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:workhours/Screens/Login/login_screen.dart';
 import 'package:workhours/Screens/Perfil/perfil.dart';
@@ -6,7 +7,11 @@ import 'package:workhours/Screens/Welcome/welcome_screen.dart';
 import 'Screens/BottomNavigator/bottomNavBar.dart';
 import 'Screens/Sobre/sobre.dart';
 
-void main() {
+Future<void> main() async {
+  //Firebase
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+
   runApp(MaterialApp(
     debugShowCheckedModeBanner: false,
     title: 'ClickPonto',
@@ -14,9 +19,9 @@ void main() {
     theme: ThemeData(
       primaryColor: Color.fromRGBO(111, 53, 165, 1),
     ),
-    initialRoute: '/welcome',
+    initialRoute: '/login',
     routes: {
-      '/welcome':(context) => WelcomeScreen(),
+      '/welcome': (context) => WelcomeScreen(),
       '/home': (context) => BottomNavBar(),
       '/sobre': (context) => Sobre(),
       '/login': (context) => LoginScreen(),
@@ -24,4 +29,3 @@ void main() {
     },
   ));
 }
-
